add mbf2fnt
This tool converts an MBF font to the following format:

"""
A font is simply stored in a directory named WxH where W and H are
the pixel  dimensions of the font. Each glyph is then stored as an
ASCII file named U+NNNN where NNNN is the Unicode codepoint of the
glyph.  Here is what is stored in  7x10/U+0041 (glyph 'A') for the
leon font:

.......
.XXX...
X...X..
X...X..
XXXXX..
X...X..
X...X..
X...X..
.......
.......
"""

This format is useful for font editing.
Reference: <https://github.com/mpu/fnt>.